% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{construct_l}
\alias{construct_l}
\title{Create L as the number of nodes}
\usage{
construct_l(sm_sparse)
}
\arguments{
\item{sm_sparse}{sparse similarity matrix object handed by the user}
}
\description{
Returns an object of class matrix for storing the hypotetical affinity scored candidate pairs for the input graphs
}
