% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{construct_x}
\alias{construct_x}
\title{Create X as the assigments with maximun confidence value}
\usage{
construct_x(sm_sparse)
}
\arguments{
\item{sm_sparse}{sparse similarity matrix object handed by the user}
}
\description{
Returns an object of class matrix for storing the highest affinity scored candidate pairs for the input graphs
}
