% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{constructClass}
\alias{constructClass}
\title{Construct sstr object}
\usage{
constructClass(smi = 0, mmi = 0, t = 0)
}
\arguments{
\item{smi}{list with input graph A sketch map with affinity candidates}

\item{mmi}{list with input graph B metric map with affinity candidates}

\item{t}{object from end function}
}
\value{
r_sstr an sstr class object
}
\description{
Return an object of class sstr
}
\examples{
start()
}
