% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageRank.R
\name{pagerankR}
\alias{pagerankR}
\title{Pagerank implementation}
\usage{
pagerankR(X, alpha = 0.85, max_iter = 50, tol = 1e-10)
}
\arguments{
\item{X}{similarity matrix of class matrix}

\item{alpha}{probability at each node to be changed}

\item{max_iter}{number of iterations for the power function}

\item{tol}{float number for convergence tolerance}
}
\description{
This pagerank coding attemps to generate the scoring for a stored similarity matrix
}
