% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{pr_scores}
\alias{pr_scores}
\title{Generates the affinity scores with the Page rank algorithm by Google from igraph}
\usage{
pr_scores(simMatrix)
}
\arguments{
\item{simMatrix}{similarity matrix object handed by the user, it should be squared and it will be treated as a direct graph}
}
\value{
score matrix for a similarity matrix
}
\description{
Returns an object of class matrix with the affinity scores
}
